Move variable creation to function
Until now, variables were created with a big if-statement
that you had to copy for every new variable. Now there's
a function that you can call. It is used as follows:

  read_variable <variable_name> <description>

The function checks if the variable is already set. If that
is the case, the user is asked if he wants to change the
value. In case it doesn't exist yet, the user is prompted
for the value.

The description is used in the prompts, and also as a
comment in user_config.sh above the line where the variable
is declared.