Add file and command asserts
This commit adds two functions that can be used by scripts
to verify things at runtime.

- assert_installed checks that a program is on the path.
  If not, it exits the current process and prints an error
  message if one was given.

- assert_value_set checks that a enironment variable has
  a value. If not, it exits the current process and prints
  an error message, if one was given.

These assertion functions can't be used in other functions
that were sourced from a file. If the assertion fails, it
would exit the current shell. That's why this commit also
adds the possibility to run scripts from the /bin folder.
The scripts in this folder can load the builtins.sh file,
where the assertions are defined, and use them from there.
Bash starts a new subprocess when running a script like
that, so the assertions will just kill the script execution
and not the entire shell.

This resolves #1.